+++ src/PeeQL.php
@@ -2,7 +2,10 @@
 
 namespace PeeQL;
 
+use Exception;
+use PeeQL\Parser\PeeQLParser;
 use PeeQL\Router\PeeQLRouter;
+use PeeQL\Schema\PeeQLSchema;
 
 /**
  * This is the main PeeQL class that has definitions of all important methods and functions
@@ -11,11 +14,16 @@ use PeeQL\Router\PeeQLRouter;
  */
 class PeeQL {
     private ?PeeQLRouter $router;
+    private ?PeeQLSchema $schema;
+    private PeeQLParser $parser;
 
     /**
      * Class constructor
      */
     public function __construct() {
+        $this->parser = new PeeQLParser();
+        
+        $this->schema = null;
         $this->router = null;
     }
 
@@ -33,12 +41,31 @@ class PeeQL {
     }
 
     /**
-     * Processes the given JSON $json query and returns the result
+     * Processes the given JSON $json query, executes it and returns the result
      * 
      * @param string $json JSON query
      */
-    public function query(string $json): mixed {
-        return null;
+    public function execute(string $json): mixed {
+        if($this->router === null) {
+            throw new Exception('No routes are defined.', 9999);
+        }
+        if($this->schema === null) {
+            throw new Exception('No schemas are defined.', 9999);
+        }
+        return $this->parser->parse($this->router, $this->schema, $json);
+    }
+
+    /**
+     * Returns an existing or a new instance of PeeQLSchema for schema definition
+     */
+    public function getSchema(): PeeQLSchema {
+        if($this->schema === null) {
+            $this->schema = new PeeQLSchema();
+        }
+
+        $schema = &$this->schema;
+
+        return $schema;
     }
 }