+++ src/schema/GetUsersSchema.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace PeeQL\Schema;
+
+class GetUsersSchema extends AQuerySchema {
+    protected function define() {
+        $this->addMultipleColumns([
+            'userId',
+            'username',
+            'fullname'
+        ]);
+    }
+}
+
+?>