+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 [PeeQL `v1.0` (_WIP_)](#peeql-v10-wip)
 
 ## PeeQL `v1.0` (_WIP_)
+- Added support for basic conditions
+- Added query operation
+- Added parser
+- Added router
+- Added query templates
+- Added support for basic schema definition